clipp: restore rapid stuff, separate narrow defs
